Kind of cleanup
Move the configuration parsing for notary-server to its own file.
